+++ client/components/DataManagement/index.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { Helmet } from 'react-helmet';
 import { Container, Table, Alert } from 'react-bootstrap';
 import { useQuery } from '@apollo/client';
@@ -14,7 +14,10 @@ import {
     DATA_MANAGEMENT_TABLE_SORT_OPTIONS
 } from '../../utils/constants';
 import FilterButtons from '../common/FilterButtons';
-import { useDataManagementTableSorting } from './hooks';
+import {
+    useDataManagementTableFiltering,
+    useDataManagementTableSorting
+} from './hooks';
 
 const DataManagement = () => {
     const { loading, data, error, refetch } = useQuery(
@@ -50,68 +53,14 @@ const DataManagement = () => {
         }
     }, [data]);
 
-    const [
-        rdTestPlans,
-        draftTestPlans,
-        candidateTestPlans,
-        recommendedTestPlans
-    ] = useMemo(() => {
-        return testPlans.reduce(
-            (acc, testPlan) => {
-                const testPlanVersion = testPlanVersions.find(
-                    ({ testPlan: { directory } }) =>
-                        directory === testPlan.directory
-                );
-                if (!testPlanVersion) return acc;
-                const phase = testPlanVersion.phase;
-                switch (phase) {
-                    case 'RD':
-                        acc[0].push(testPlan);
-                        break;
-                    case 'DRAFT':
-                        acc[1].push(testPlan);
-                        break;
-                    case 'CANDIDATE':
-                        acc[2].push(testPlan);
-                        break;
-                    case 'RECOMMENDED':
-                        acc[3].push(testPlan);
-                        break;
-                    default:
-                        break;
-                }
-                return acc;
-            },
-            [[], [], [], []]
-        );
-    }, [testPlans]);
-
-    const filteredTestPlans = useMemo(() => {
-        switch (filter) {
-            case DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.RD:
-                return rdTestPlans;
-            case DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.DRAFT:
-                return draftTestPlans;
-            case DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.CANDIDATE:
-                return candidateTestPlans;
-            case DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.RECOMMENDED:
-                return recommendedTestPlans;
-            case DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.ALL:
-            default:
-                return testPlans;
-        }
-    }, [filter, testPlans]);
+    const { filteredTestPlans, filterLabels } = useDataManagementTableFiltering(
+        testPlans,
+        testPlanVersions,
+        filter
+    );
 
     const { sortedTestPlans, updateSort, activeSort } =
         useDataManagementTableSorting(filteredTestPlans, testPlanVersions, ats);
-
-    const filterLabels = {
-        [DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.RD]: `R&D Complete (${rdTestPlans.length})`,
-        [DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.DRAFT]: `In Draft Review (${draftTestPlans.length})`,
-        [DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.CANDIDATE]: `In Candidate Review (${candidateTestPlans.length})`,
-        [DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.RECOMMENDED]: `Recommended Plans (${recommendedTestPlans.length})`,
-        [DATA_MANAGEMENT_TABLE_FILTER_OPTIONS.ALL]: `All Plans (${testPlans.length})`
-    };
 
     if (error) {
         return (
